Update jquery.viewportchecker.js
In this plugin, jQuery keeps triggering events after the $elem is removed from the DOM. This causes a speed problem if the plugin is applied multiple times (after loading new content via ajax). After some time, the instances of the plugin are looping over a lot of removed elements. 

I have therefore added the variable "namespace" and function "destroy" to handle this, and it seems to work pretty well.